requirements.txt
asgiref==3.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
dj-database-url==3.0.1
Django==6.0
django-cors-headers==4.9.0
django-decouple==2.1
djangorestframework==3.16.1
idna==3.10
psycopg==3.3.2
psycopg-binary==3.3.2
requests==2.32.5
sqlparse==0.5.5
urllib3==2.5.0
whitenoise==6.11.0
